tools reorganization
