+++ shared/shell_tools.py
